chore: remove receipt mocking from upload-client tests
